// src/app/shared/types/article.interface.ts
import {ProfileInterface} from 'src/app/shared/types/profile.interface'

export interface ArticleInterface {
  author: ProfileInterface
  body: string
  createdAt: string
  description: string
  favorited: boolean
  favoritesCount: number
  slug: string
  tagList: string[]
  title: string
  updatedAt: string
}